refactor: 북마크, Recommendation Converter 리팩토링
