+++ src/main/java/com/eodya/api/bookmark/domain/BookMarkStatus.java
@@ -1,7 +1,6 @@
 package com.eodya.api.bookmark.domain;
 
 import com.eodya.api.bookmark.exception.BookMarkException;
-import com.eodya.api.recommendation.exception.RecommendationException;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 import lombok.Getter;
@@ -14,7 +13,6 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static com.eodya.api.bookmark.exception.BookMarkExceptionCode.*;
-
 
 @Getter
 @RequiredArgsConstructor
@@ -32,7 +30,7 @@ public enum BookMarkStatus {
         if (statusMap.containsKey(description)) {
             return statusMap.get(description);
         }
-        throw new RecommendationException(BOOKMARK_STATUS_NOT_FOUND, description);
+        throw new BookMarkException(BOOKMARK_STATUS_NOT_FOUND, description);
     }
 
     private static final Map<String, BookMarkStatus> statusMap =
