+++ src/main/java/com/eodya/api/recommendation/domain/Recommendation.java
@@ -2,7 +2,7 @@ package com.eodya.api.recommendation.domain;
 
 import com.eodya.api.common.entity.BaseEntity;
 import com.eodya.api.place.domain.Place;
-import com.eodya.api.recommendation.util.RecommendationStatusAttributeConverter;
+import com.eodya.api.recommendation.util.RecommendationStatusConverter;
 import com.eodya.api.users.domain.User;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
@@ -27,7 +27,7 @@ public class Recommendation extends BaseEntity {
     private Long id;
 
     @NotNull
-    @Convert(converter = RecommendationStatusAttributeConverter.class)
+    @Convert(converter = RecommendationStatusConverter.class)
     @Column(length = 20)
     private RecommendationStatus status = FALSE;
