--- src/main/java/com/eodya/api/recommendation/util/RecommendationStatusAttributeConverter.java
@@ -1,19 +1,0 @@
-package com.eodya.api.recommendation.util;
-
-import com.eodya.api.recommendation.domain.RecommendationStatus;
-import jakarta.persistence.AttributeConverter;
-import jakarta.persistence.Converter;
-
-@Converter
-public class RecommendationStatusAttributeConverter implements AttributeConverter<RecommendationStatus, String> {
-
-    @Override
-    public String convertToDatabaseColumn(RecommendationStatus recommendationStatus) {
-        return recommendationStatus.getDescription();
-    }
-
-    @Override
-    public RecommendationStatus convertToEntityAttribute(String status) {
-        return RecommendationStatus.from(status);
-    }
-}
